Reload whole form with ajax
