+++ app/controllers/challenges_controller.rb
@@ -61,9 +61,15 @@ class ChallengesController < ApplicationController
 
   def update
     if @challenge.update(challenge_params)
-      redirect_to edit_challenge_path(@challenge, step: params[:next_step]), notice: 'Challenge updated.'
+      respond_to do |format|
+        format.html { redirect_to edit_challenge_path(@challenge, step: params[:next_step]), notice: 'Challenge updated.' }
+        format.js   { render :update }
+      end
     else
-      render :edit
+      respond_to do |format|
+        format.html { render :edit }
+        format.js   { render :update }
+      end
     end
   end
